Be reliable about the type of EmarsysError.code
Since code can be a string or an int, you'd usually save it as an int in
some kind of log. For Django ORM records, the value is coerced to a string
only when it's saved to the database, and then subsequently retrieved. As
long as you're using the same fresh object you initially assigned the code
to, it's still an int. So if you want to check for error code 2008, you
have to check for the int 2008 if you use the fresh object, and for the
string '2008' if it's an object retrieved from the database. Or you try to
convert to an int and have to handle ValueError. All that's a code smell.